### 2020-11-10 #2833 @kthyng
Hi @salrm8! Can you look through and address the DOI problems above and look through the errors with building your paper to try to fix them?
