+++ paper/paper.bib
@@ -9,7 +9,7 @@
 
 @book{Santner:2003,
      doi = {10.1007/978-1-4757-3799-8},
-     Adsurl = {https://doi.org/10.1007/978-1-4757-3799-8},
+     Adsurl = {https://link.springer.com/book/10.1007/978-1-4757-3799-8},
      year = 2003,
      Publisher = {Springer New York},
      Author = {Thomas J. Santner and Brian J. Williams and William I. Notz},
@@ -31,6 +31,7 @@
     Publisher = {Chapman Hall/CRC},
     Address = {Boca Raton, Florida},
     Adsurl = {http://bobby.gramacy.com/surrogates/},
+    isbn={9781000766523},
     Year = 2020
 }
 
@@ -42,6 +43,7 @@
    number = {2},
    pages = {619--644},
    year = 2002,
+   doi={10.1137/S1064827501387826},
    Adsurl={https://epubs.siam.org/doi/10.1137/S1064827501387826}
 }
 
@@ -65,7 +67,7 @@
    Year = 2001,
    Adsnote = "The Second IMACS Seminar on Monte Carlo Methods",
    issn = "0378-4754",
-   doi = "https://doi.org/10.1016/S0378-4754(00)00270-6",
+   doi = "10.1016/S0378-4754(00)00270-6",
    Adsurl = "http://www.sciencedirect.com/science/article/pii/S0378475400002706",
    Author = "I.M Sobol",
 }
@@ -79,7 +81,6 @@
    Number="",
    Pages="",
    issn="",
-   doi="arXiv:2007.07071",
    Adsurl="https://arxiv.org/abs/2007.07071"
 }
 
@@ -96,17 +97,18 @@
    Author = {R. C. Smith},
    Title = {Uncertainty Quantification: Theory, Implementation, and Applications},
    Year = 2013,
-   isbn = {161197321X, 9781611973211},
+   isbn = {9781611973211},
    Publisher = {Society for Industrial and Applied Mathematics},
    Address = {Philadelphia, PA, USA},
-   Adsurl={https://rsmith.math.ncsu.edu/UQ_TIA/}
+   Adsurl={https://my.siam.org/Store/Product/viewproduct/?ProductId=24973024}
 }
 
 @book{Sagaut:2013,
   Title={Multiscale and Multiresolution Approaches in Turbulence: {LES}, {DES} and Hybrid {RANS}/{LES} Methods: Applications and Guidelines},
   Author={Pierre Sagaut and Sebastian Deck and Marc Terracol},
   isbn={9781848169876},
-  Adsurl={https://books.google.se/books?id=MzW6CgAAQBAJ},
+  Adsurl={https://www.worldscientific.com/doi/abs/10.1142/p878},
+  doi={10.1142/p878},
   Year=2013,
   Publisher={Imperial College Press}
 }
