+++ apps/turbChannel_CE/modules/nekProfsReader_tChan.py
@@ -1,341 +1,0 @@
-####################################################################
-# Read in  and construct databases from 
-#      the statistics profiles of turbulent channel flow
-####################################################################
-#   Saleh Rezaeiravesh, salehr@kth.se
-#-------------------------------------------------------------------
-import sys
-import os
-import numpy as np
-import math as mt
-import matplotlib.pyplot as plt
-sys.path.append(os.getenv("UQit"))
-from lagInt import lagInt
-#
-def lagrangeInterpGLL2UnifMesh(y,f,interpOpts):
-    """
-    Lagrange interpolation from GLL points at each NEK5000 element to an arbitrary number of uniformly-spaced points inside the same element. 
-    f(y) is mapped into fInt(yInt)
-
-    Args:
-      `y`: 1D numpy array 
-         Containing the global coordinates of the GLL points
-      `f`: 1D numpy array
-         Values at `y`
-      `interpOpts`: dict   
-         Options for interpolation
-    
-    Returns:
-      `yInt`: 1D numpy array
-         Containing the global coordinates of yInt
-      `fInt`: 1D numpy array
-         Interpolated f at `yInt`
-    """
-    nGLL=interpOpts['nGLL']   
-    nTot=len(y)                 
-    nElement=int(nTot/nGLL)    
-    nIntPerE=interpOpts['nIntPerE'] 
-
-    yGLL=np.zeros(nGLL)   
-    fGLL=np.zeros(nGLL)
-    yInt=[]
-    fInt=[]
-    for i in range(nElement):
-        for j in range(nGLL):
-            k=i*nGLL+j 
-            yGLL[j]=y[k]               
-            fGLL[j]=f[k]               
-        yInt_E=np.linspace(np.min(yGLL),np.max(yGLL),nIntPerE)
-        fInt_E=lagInt(fNodes=fGLL,qNodes=[yGLL],qTest=[yInt_E]).val
-        for j in range(nIntPerE):
-            yInt.append(yInt_E[j])
-            fInt.append(fInt_E[j])
-    yInt=np.asarray(yInt)
-    fInt=np.asarray(fInt)
-    return yInt,fInt
-#
-def sortedCaseInfo(caseDir,caseName):
-    """ 
-    Read information about a Nek5000 case (a set of channel flow simulations)
-    The info is saved in a database which will be sorted to make sure our convention for tensor-product in parameter space holds. This sorted info-db will be later used to sort other databases of the case.
-    The info file contains information about uncertain parameters associated with each case and has the following format.
-
-      # Information on the simlations set <caseName>. \n
-      # caseName: \n
-      # Number of included simulations: \n
-      # Number of uncertain parameters: \n
-      # Number of parameter samples in each direction:  (non-tensor-product: nTotSamples 1 1 ... 1) \n
-      # Name of uncertain parameters: \n
-      # Admissible range of parameters: [q1Min q1Max] [q2Min q2Max] \n
-      #List of simulations name and associated parameter values \n
-      # simName    q1      q2 \n
-      #--------------------------------------------------------- (iEnd1)
-    
-    Args:
-      `caseDir`: string
-         Path to the Nek5000 simulations data
-      `caseName`: string
-         Name of the simulation case
-    
-    Returns:
-      `db_info`: dict
-         Information about the Nek5000 set of simulations
-    """   
[…]
