### 2020-11-23 #2833 @drvinceknight
@salrm8 do you have any suggestions for reviewer? Looking through the list of potential reviewers: https://bit.ly/joss-reviewers a number jump out as having experience with both Python and fluid dynamics. No trouble at all if you don't have any suggestions :)
