### 2020-11-23 #2833 @salrm8
> @salrm8 do you have any suggestions for reviewer? Looking through the list of potential reviewers: https://bit.ly/joss-reviewers a number jump out as having experience with both Python and fluid dynamics. No trouble at all if you don't have any suggestions :)

Thanks you. I just did a quick search and came up with this list: prmiles, georgiastuart, ctdegroot, jayten
They have experience with UQ, Python and CFD.
