### 2020-11-23 #2833 @drvinceknight
Hello @jayten and @ctdegroot, would you be able to review this paper for JOSS entitled: "UQit: A Python package for uncertainty quantification (UQ) in computational fluid dynamics". Given your stated expertise in fluid dynamics and Python I think you'll find it interesting :)

If either of you have any questions about reviewing for JOSS or anything else let me know. There is also very helpful documentation here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#.
