### 2020-11-25 #2833 @ctdegroot
Yep, I am interested in this. Might take me a couple weeks to get to it as I'm wrapping up another JOSS review.
