### 2020-11-30 #2871 @whedon
[REVIEW]: UQit: A Python package for uncertainty quantification (UQ) in computational fluid dynamics (CFD)
**Submitting author:** @salrm8 (<a href="http://orcid.org/0000-0002-9610-9910">Saleh Rezaeiravesh</a>)
**Repository:** <a href="https://github.com/KTH-Nek5000/UQit" target ="_blank">https://github.com/KTH-Nek5000/UQit</a>
**Version:** v1.0.2
