### 2020-12-11 #2871 @jayten
### Overview
The submitted software is well-written and acceptable for publication in JOSS.

### Regarding authors
The actual name of the contributor "saleh8" is not clear. If he/she is not the already existing author, consider adding his/her name to the list of authors or in the acknowledgment, accordingly.

### Documentation
It will be better for people like me who do not use "conda" to have an already existing Documentation website or a PDF file. Already existing HTML files can be easily hosted on GitHub using [Github pages](https://pages.github.com/).

### Community guidelines
Since the authors plan to add a few more features in the future release, consider adding a simple contriubting.txt file to guide people interested in helping you with that work.

### Test

- Is it possible to automate the tests?

- **Error encountered while running tests**:

1. When I am trying to run 2d and 3d PCE tests, it am getting the following error:
```
Traceback (most recent call last):
  File "pce_tests.py", line 301, in <module>
    pce_2d_test()
  File "pce_tests.py", line 153, in pce_2d_test
    pce_=pce(fVal=fVal,xi=xiGrid,pceDict=pceDict,nQList=nQ)
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/UQit/pce.py", line 92, in __init__
    self.cnstrct()
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/UQit/pce.py", line 343, in cnstrct
    self.cnstrct_pd()
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/UQit/pce.py", line 446, in cnstrct_pd
    self.cnstrct_nonGQTP_pd()
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/UQit/pce.py", line 595, in cnstrct_nonGQTP_pd
    fCoef=linAlg.myLinearRegress(A,self.fVal)
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/UQit/linAlg.py", line 53, in myLinearRegress
    f=linearSysSolver(A,R,1)
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/UQit/linAlg.py", line 46, in linearSysSolver
    object_value = prob.solve(verbose=True,max_iter=max_iter_)
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/cvxpy/problems/problem.py", line 396, in solve
    return solve_func(self, *args, **kwargs)
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/cvxpy/problems/problem.py", line 752, in _solve
    solution = solving_chain.solve_via_data(
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/cvxpy/reductions/solvers/solving_chain.py", line 325, in solve_via_data
    return self.solver.solve_via_data(data, warm_start, verbose,
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/cvxpy/reductions/solvers/conic_solvers/ecos_conif.py", line 137, in solve_via_data
    solution = ecos.solve(data[s.C], data[s.G], data[s.H],
  File "/home/jatinder/anaconda3/lib/python3.8/site-packages/ecos/ecos.py", line 62, in solve
    return _ecos.csolve((m,n1,p), c, data, indices, colptr, h, dims, A.data, A.indices, A.indptr, b, **kwargs)
TypeError: 'max_iter' is an invalid keyword argument for this function
```

2. Figures expected by the test for Lagrange interpolation are missing. The following error was issued:
`FileNotFoundError: [Errno 2] No such file or directory: '../testFigs/lag1d4.png'`
