### 2020-12-14 c73790a
default solver_ added; different solvers can be used
