### 2020-12-14 abbccea
updating linAlg doc adding notes about solver
