### 2020-12-14 699f0e4
docs/source/* moved to docs/*
