+++ docs/conf.py
@@ -12,7 +12,7 @@
 #
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../UQit/'))
+sys.path.insert(0, os.path.abspath('../UQit/'))
 ##sys.path.insert(1, os.path.abspath('../../apps/turbChannel_CE/modules/'))
 import sphinx_rtd_theme
 
@@ -23,7 +23,7 @@ copyright = '2020, Saleh Rezaeiravesh'
 author = 'Saleh Rezaeiravesh'
 
 # The full version, including alpha/beta/rc tags
-release = '1.0.1'
+release = '1.0.3'
 
 # -- General configuration ---------------------------------------------------
 
