+++ docs/Makefile
@@ -5,8 +5,10 @@
 # from the environment for the first two.
 SPHINXOPTS    ?=
 SPHINXBUILD   ?= sphinx-build
-SOURCEDIR     = source
-BUILDDIR      = build
+#SOURCEDIR     = source
+#BUILDDIR      = build
+SOURCEDIR     = ./
+BUILDDIR      = _build
 
 # Put it first so that "make" without argument is like "make help".
 help:
