### 2020-12-14 76eef4b
log path modified due to docs/
