+++ docs_build/_sources/examples/sobol.ipynb.txt
@@ -1,0 +1,818 @@
+{
+ "cells": [
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "# Sobol Sensitivity Indices\n",
+    "\n",
+    "Saleh Rezaeiravesh, salehr@kth.se <br/>\n",
+    "SimEx/FLOW, Engineering Mechanics, KTH Royal Institute of Technology, Stockholm, Sweden\n"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 1,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "import os\n",
+    "import sys\n",
+    "import numpy as np\n",
+    "import math as mt\n",
+    "import matplotlib.pyplot as plt\n",
+    "from UQit.sobol import sobol\n",
+    "import UQit.analyticTestFuncs as analyticTestFuncs\n",
+    "from UQit.pce import pce, pceEval\n",
+    "import UQit.reshaper as reshaper\n",
+    "import UQit.sampling as sampling"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "### Example 1: Sobol indices for 2 uniform parameters"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "Consider the following analytical function, \n",
+    "$$\n",
+    "\\begin{equation}\n",
+    "f(\\mathbf{q}) = q_1^2+q_1 q_2\n",
+    "\\label{eq:1dModel}\\tag{3}\n",
+    "\\end{equation}\n",
+    "$$\n",
+    "\n",
+    "where $\\mathbf{q}=\\{q_1,q_2\\}$, and both parameters are random with uniform distributions, i.e., $q_1\\sim \\mathcal{U}[\\mathbb{Q}_1]$ and $q_2\\sim \\mathcal{U}[\\mathbb{Q}_2]$. \n",
+    "The aim is to compute the main Sobol indices using `UQit` and validate them with the analytical values..\n",
+    "\n",
+    "The above model function (simulator) is implemented in `analyticTestFuncs.fEx2D` with keyword `type3`. "
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "**Step 1:** Set the number of samples for $q_1$ and $q_2$ and their associated admissible spaces $\\mathbb{Q}_1$ and $\\mathbb{Q}_2$"
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 2,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "n=[100, 90]       #number of samples for q1 and q2\n",
+    "qBound=[[-3,1],   #admissible space of q1 and q2\n",
+    "        [-1,2]]"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
+    "**Step 2:** Generate samples for parameters $q_1$ and $q_2$. We choose to have samples to be uniformly spaced over the parameter spaces. "
+   ]
+  },
+  {
+   "cell_type": "code",
+   "execution_count": 3,
+   "metadata": {},
+   "outputs": [],
+   "source": [
+    "p=len(n)\n",
+    "q=[]\n",
+    "pdf=[]\n",
+    "for i in range(p):\n",
+    "    q.append(np.linspace(qBound[i][0],qBound[i][1],n[i]))\n",
+    "    pdf.append(np.ones(n[i])/(qBound[i][1]-qBound[i][0]))"
+   ]
+  },
+  {
+   "cell_type": "markdown",
+   "metadata": {},
+   "source": [
[…]
