+++ docs/UQit_core_/uqFWD.rst
@@ -1,131 +1,0 @@
-.. _uqFwd-sect:
-
-==================
-UQ Forward Problem
-==================
-
-In UQ forward (uncertainty propagation) problems, the aim is to estimate the propagation of uncertainties in the model response or QoIs, from known uncertain parameters/inputs. 
-In many situations, we are mostly interested in approximately estimating the statistical moments of the model outputs and QoIs. 
-In particular, among the moments, our main focus is on the expected value :math:`\mathbb{E}_\mathbf{q}[r]` and variance :math:`\mathbb{V}_\mathbf{q}[r]` of a QoI :math:`r`, with :math:`\mathbf{q}` denoting the uncertain parameters.
-Different approaches can be used for this purpose, see [Smith13]_, [Ghanem17]_.
-The Monte Carlo approaches rely on taking independent samples from the parameters and running the model simulator as a blackbox. 
-However, the convergence rate of these methods can be low, see e.g. [Ghanem17]_.
-As a more efficient technique, the spectral-based method non-intrusive generalized polynomial chaos expansion (gPCE) [Xiu02]_, [Xiu05]_, [Xiu07]_ is employed in :code:`UQit` for UQ forward problems. 
-For an overview of the technique with the same notations as in this document, refer to [Rezaeiravesh20]_.
-
-.. _gPCE-sect:
-
-Standard Polynomial Chaos Expansion
------------------------------------
-
-Theory
-~~~~~~
-The generalized polynomial chaos expansion for :math:`\tilde{f}(\chi,\mathbf{q})` is written as,
-
-.. math::
-   \tilde{f}(\chi,\mathbf{q}) = \sum_{k=0}^K \hat{f}_k(\chi) \Psi_{k}(\xi) \,,
-
-where the basis function for the multi-variate parameter :math:`\xi\in\Gamma` is defined as :math:`\Psi_{k}(\xi)=\prod_{i=1}^p \psi_{k_i}(\xi_i)`.
-In the framework of generalized PCE, see [Xiu02]_, given the distribution of the single-variate parameter :math:`\xi_i\in \Gamma_i`, a set of orthogonal basis functions are chosen for :math:`\psi_{k_i}(\xi_i)`, for :math:`i=1,2,\ldots,p`. 
-Note that, there is a one-to-one correspondence between any sample of :math:`\mathbf{q}\in \mathbb{Q}` and :math:`\xi\in\Gamma`, where :math:`\mathbb{Q}=\bigotimes_{i=1}^p \mathbb{Q}_i` and :math:`\Gamma=\bigotimes_{i=1}^p \Gamma_i`.
-The mapped space :math:`\Gamma_i` is known based on the gPCE rule, see [Xiu02]_, [Eldred09]_.
-
-
-Given a set of training data :math:`\mathcal{D}=\{(\mathbf{q}^{(i)},r^{(i)})\}_{i=1}^n`, there are two main steps to construct the above expansion.
-First, a truncation scheme is needed to handle :math:`p`-dimensional parameter space and determine :math:`K`.
-Currently, tensor-product and total-order schemes are available in :code:`UQit`. 
-Second, the coefficients :math:`\{\hat{f}_k(\chi)\}_{k=0}^K` have to be determined. 
-In :code:`UQit`, two different approaches can be used for this purpose: projection and regression method, see [Rezaeiravesh20]_ and the references therein. 
[…]

[…]
