+++ docs/bib.rst
@@ -1,0 +1,40 @@
+============
+Bibliography
+============
+
+.. [Smith13] `R. C. Smith. Uncertainty Quantification: Theory, Implementation, and Applications. Society for Industrial and Applied Mathematics, Philadelphia, PA, USA, 2013. <https://rsmith.math.ncsu.edu/UQ_TIA/>`_
+
+.. [Ghanem17] `R. Ghanem, D. Higdon, and H. Owhadi, editors. Handbook of Uncertainty Quantification. Springer International Publishing, 2017. <https://www.springer.com/gp/book/9783319123844>`_
+
+.. [Xiu02] `D. Xiu and G. E. Karniadakis. The Wiener–Askey polynomial chaos for stochastic differential equations. SIAM Journal on Scientific Computing, 24(2):619–644, 2002. <https://epubs.siam.org/doi/10.1137/S1064827501387826>`_
+
+.. [Xiu05] `D. Xiu and J. S. Hesthaven. High-order collocation methods for differential equations with random inputs. SIAM Journal on Scientific Computing, 27(3):1118–1139, 2005. <https://epubs.siam.org/doi/10.1137/040615201>`_
+
+.. [Xiu07] `D. Xiu. Efficient Collocational Approach for ParametricUncertainty Analysis. Communications in Computational Physics, 2(2): 293-309, 2007 <http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.324.2923&rep=rep1&type=pdf>`_
+
+.. [Rezaeiravesh18] `S. Rezaeiravesh, R. Vinuesa, M. Liefvendahl, and P. Schlatter. Assessment of uncertainties in hot-wire anemometry and oil-film interferometry measurements for wall-bounded turbulent flows. European Journal of Mechanics - B/Fluids, 72:57-73, 2018. <https://www.sciencedirect.com/science/article/abs/pii/S099775461730496X>`_
+
+.. [Rezaeiravesh20] `S. Rezaeiravesh, R. Vinuesa and P. Schlatter, An Uncertainty-Quantification Framework for Assessing Accuracy, Sensitivity, and Robustness in Computational Fluid Dynamics, arXiv:2007.07071, 2020. <https://arxiv.org/abs/2007.07071>`_
+
+.. [Diamond16] `S. Diamond and S. Boyd. CVXPY: A Python-embedded modeling language for convex optimization. Journal of Machine Learning Research, 17(83):1{5, 2016. <https://www.cvxpy.org/index.html>`_
+
+.. [Rasmussen05] `C. E. Rasmussen and C. K. I. Williams. Gaussian Processes for Machine Learning (Adaptive Computation and Machine Learning). The MIT Press, 2005. ISBN 026218253X. <http://www.gaussianprocess.org/gpml/>`_
+
+.. [Goldberg98] `. W. Goldberg, C. K. I. Williams, and C. M. Bishop. Regression with input-dependent noise: A gaussian process treatment. In Proceedings of the 1997 Conference on Advances in Neural Information Processing Systems 10, NIPS 97, page 493-499, Cambridge, MA, USA, 1998. MIT Press. ISBN 0262100762. <https://www.microsoft.com/en-us/research/publication/regression-with-input-dependent-noise-a-gaussian-process-treatment/>`_
+
+.. [Sobol01] `Sobol, I. Global sensitivity indices for nonlinear mathematical models and their monte carlo estimates. Mathematics and Computers in Simulation, 55(1):271 – 280, 2001. <https://www.sciencedirect.com/science/article/abs/pii/S0378475400002706>`_
+
[…]
