+++ docs/_sources/UQit_core_/codes_list.rst.txt
@@ -1,0 +1,30 @@
+==================
+List of core codes
+==================
+
+* :code:`analyticTestFuncs.py`: 
+  Analytical model functions to test implementation of different techniques
+* :code:`gpr_torch.py`:
+  Gaussian Process Regression (GPR) using GPyTorch library
+* :code:`lagInt.py`:
+  Lagrange interpolation
+* :code:`linAlg.py`:
+  Tools for linear algebra 
+* :code:`nodes.py`:
+  Spectral nodes from different types of polynomials
+* :code:`pce.py`: 
+  generalized Polynomial Chaos Expansion
+* :code:`stats.py`:
+  Statistical tools
+* :code:`ppce.py`:
+  Probabilistic generalized Polynomial Chaos Expansion (PPCE)
+* :code:`reshaper.py`:
+  Tools for converting and reshaping arrays and lists  
+* :code:`sampling.py`: 
+  Sampling from parameter spaces
+* :code:`sobol.py`:
+  Sobol sensitivity indices
+* :code:`surr2surr.py`:
+  Interpolation from a surrogate to another surrogate
+* :code:`writeUQ.py`:  
+  Tools for printing or writing data in file
