+++ docs/_sources/UQit_core_/surrogate.rst.txt
@@ -1,0 +1,134 @@
+.. _surrogates-sect:
+
+==========
+Surrogates
+==========
+
+A surrogate or metamodel is an approximation of the actual model function or simulator over the parameter space.
+Running a surrogate is computationally much less expensive than the actual simulator, a characteristic that makes the use of surrogates inevitable in different UQ problems.
+However, the predictions by the surrogate should be accurate enough compared to the actual predictions by the simulator. 
+Using an additive error model, the following relation can be considered between the model function (simulator) and its observations:
+
+.. math::
+   r=f(\chi,\mathbf{q})+\mathbf{\varepsilon}\,,
+
+where :math:`\mathbf{\varepsilon}` expresses the bias and random noise.
+Our task is to construct a surrogate :math:`\tilde{f}(\chi,\mathbf{q})` for the actual model function (simulator).
+
+Treating the simulator as a blackbox, a set of training data :math:`\mathcal{D}=\{(\mathbf{q}^{(i)},r^{(i)})\}_{i=1}^n` is obtained. 
+There are different techniques to construct a surrogate, for instance see [Smith13]_, [Ghanem17]_, and [Gramacy20]_. 
+Some of the techniques relevant to CFD applications have been implemented in :code:`UQit`.
+Here, we provide a short overview to the theory behind these techniques, explain their implementation in :code:`UQit`, and provide examples to show how to use them. 
+
+
+Non-intrusive Polynomial Chaos Expansion
+-----------------------------------------
+As a type of stochastic collocation (SC) methods, see e.g. Chapter 20 in [Ghanem17]_, non-intrusive PCE [Xiu05]_, [Xiu07]_ can be used to construct a surrogate. 
+
+.. math::
+   \tilde{f}(\chi,\mathbf{q}) = \sum_{k=0}^K \hat{f}_k(\chi) \Psi_{k}(\xi) \,.
+
+There is a one-to-one correspondence between any sample of :math:`\mathbf{q}\in \mathbb{Q}` and :math:`\xi\in\Gamma`, where :math:`\mathbb{Q}=\bigotimes_{i=1}^p \mathbb{Q}_i` and :math:`\Gamma=\bigotimes_{i=1}^p \Gamma_i`. 
+Note that :math:`\mathbb{Q}_i` is the admissible space of the i-th parameter which can be mappd onto :math:`\Gamma_i` based on the gPCE rules, see [Xiu02]_, [Eldred09]_.
+For the details of the non-intrusive PCE method refer to :ref:`gPCE-sect`.
+
+
+
+Lagrange Interpolation
+----------------------
+
+Theory
+~~~~~~
+As another form of SC-based surrogates, Lagrange interpolation can be considered:
+
+.. math::
+   \tilde{f}(\chi,\mathbf{q}) = \sum_{k=1}^n \hat{f}_k(\chi,\mathbf{q}) L_k(\mathbf{q}) \,,
+
+where :math:`\hat{f}_k(\chi,\mathbf{q})=f(\chi,\mathbf{q}^{(k)})=r^{(k)}` are the training model outputs.
+If the :math:`n_i` samples taken from the :math:`i`-th parameter space are represented by :math:`Q_{n_i}`, then the use of tensor-product leads to the nodal set :math:`Q_n` of size :math:`n=\prod_{i=1}^p n_i`, where,
+
+.. math::
+   Q_n= Q_{n_1} \bigotimes Q_{n_2}\bigotimes \ldots \bigotimes Q_{n_p} \,.
+
[…]
