### 2020-12-14 aee9db0
docs->docsrc for the logo
