+++ CONTRIBUTING.md
@@ -1,0 +1,6 @@
+`UQit` is planned to be continuously developed.
+Therefore, any contribution in the code development and documentation is warmly welcomed!
+
+* In case you are interested, please contact the authors listed in `README.md`.
+
+* A general guide for how to contribute in an open-source code is found in this [GitHub link](https://github.com/github/docs/blob/main/CONTRIBUTING.md).
