### 2020-12-15 #2871 @salrm8
@drvinceknight and @jayten: Thank you again for the comments. Small modifications in the codes and documentation are applied and pushed to the repo.
The summary of the actions/answers are as follows:

1. salrm8 and saleh8 are the IDs of the first author in the JOSS paper.
2. A GitHub page is created to host the documentation: https://kth-nek5000.github.io/UQit/
3. File "CONTRIBUTING.md" is created to facilitate contributions.
4. Run-time error 1: The 2d and 3d PCE tests should run now. The issue was that the setting "maxiter" in the linear solver in linAlg.py was set based on the default solver OSQP. Now different linear solvers required by [cxypy](https://www.cvxpy.org/tutorial/advanced/index.html?highlight=installed_solvers) can be used.
5. Run-time error 2: Figure '../testFigs/lag1d4.png' was an output of the test. Now, saving the figure has been removed from the test.
6. The only remaining comment is about automating the tests. Since running each test generates a set of plots, we thought it might be unnecessary to automate them. Also, the tests have also been used in the notebooks in the documentation.

I hope these could answer the comments.
