### 2020-12-15 #2871 @salrm8
@jayten Has UQit been re-installed? It seems the previous version is still being used. One can simply replace "/home/jatinder/anaconda3/lib/python3.8/site-packages/UQit/linAlg.py" by the linAlg.py from the repo.

After this test, I will upload a new version of UQit on PyPI.
Thank you,
