+++ UQit/gpr_torch.py
@@ -21,6 +21,14 @@ Notes:
   parameter space can be mapped, for instance, into the hypercube
   [-1,1]^p. Then, the GPR can be constructed on this mapped space. 
 
+  4. Shifting and scaling (standardization) of the training data can result 
+  in a better fit of GPR. This can be activated via making `'standardizeYTrain':True` 
+  in `gprOpts`.
+
+  5. Different options for kernel function are available by :code:`GPyTorch`. 
+  One can change the default kernel in `self.covar_module(...)` (see below) 
+  manually by modifying the source code of :code:`UQit`. 
+
 """    
 #----------------------------------------------------------------
 #
@@ -67,9 +75,12 @@ class SingletaskGPModel_mIn(gpytorch.models.ExactGP):
         num_dims = train_x.size(-1)
         self.mean_module = gpytorch.means.ConstantMean()
         self.covar_module = gpytorch.kernels.ScaleKernel(
+            ##different length scales in different dimentions
             #gpytorch.kernels.RBFKernel(ard_num_dims=num_dims)   
-            ##different length scales in different dimentions, RBF
-            gpytorch.kernels.MaternKernel(nu=2.5,ard_num_dims=num_dims)   
+            gpytorch.kernels.RBFKernel(ard_num_dims=num_dims,
+                                      lengthscale_constraint=None,
+                                      lengthscale_prior=gpytorch.priors.NormalPrior(10, 10)) #prior (optional), to make a better fit
+            #gpytorch.kernels.MaternKernel(nu=2.5,ard_num_dims=num_dims)   
             ##different length scales in different dimentions, Matern nu
             #gpytorch.kernels.RBFKernel()   #equal length scales in all input dimensions
         )
@@ -109,6 +120,11 @@ class gpr:
                Number of iterations in the optimization of hyperparameters
            * 'lr': float 
                Learning rate in the optimization of hyperparameters
+           * 'standardizeYTrain': bool (optional, default:False)
+              If true, the training data are standardized by shifting by mean and scaling by sdev:
+
+              :math:`yStnd =  (yTrain-mean(yTrain))/sdev(yTrain)`
+
            * 'convPlot': bool 
                If true, optimized values of the hyper-parameters is plotted vs. iteration.
 
@@ -116,6 +132,8 @@ class gpr:
       `post_f`: Posterior of f(x) at `xTest`.
       
       `post_obs`: Predictive posterior (likelihood) at `xTest`.
+
+      
     """
     def __init__(self,xTrain,yTrain,noiseV,xTest,gprOpts):
         self.xTrain=xTrain
@@ -129,6 +147,15 @@ class gpr:
     def _info(self):
        self.p=self.xTrain.shape[-1]
        self.nResp=self.yTrain.shape[-1]
+       self.shift=np.zeros(self.nResp) 
+       self.scale=np.ones(self.nResp)
+       if 'standardizeYTrain' in self.gprOpts.keys(): 
+          if self.gprOpts['standardizeYTrain']:
+             for i in range(self.nResp): 
+                 self.shift[i]=np.mean(self.yTrain[:,i])
+                 self.scale[i]=np.std(self.yTrain[:,i])
+                 self.yTrain[:,i]=(self.yTrain[:,i]-self.shift[i])/self.scale[i]
+                 print('...... GPR: y-Training data is standardized by shifting and scaling.')
 
     def train_pred(self):    
         """
@@ -200,7 +227,7 @@ class gpr:
            optimizer.step()
            losses.append(loss.item())
[…]
