+++ UQit/ppce.py
@@ -4,9 +4,6 @@
 #--------------------------------------------------------------
 # Saleh Rezaeiravesh, salehr@kth.se
 #--------------------------------------------------------------
-#TODO:
-#2. Add other distType for the pce
-#--------------------------------------------------------------
 #
 import os
 import sys
