### 2021-01-15 8778d5d
forcing model and likelihood to be double()
