### 2021-01-15 29a9569
docstring about `f` is corrected
