+++ UQit/pSobol.py
@@ -1,0 +1,424 @@
+###############################################################
+# Probabilistic Sobol Sensitivity Indices
+###############################################################
+#--------------------------------------------------------------
+# Saleh Rezaeiravesh, salehr@kth.se
+#--------------------------------------------------------------
+# ToDo: For p>=3, choosing large number of test points to evaluate integrals in Sobol indices, 
+#       is problematic since associated samples are generated from GPR (GpyTorch has memory issues). 
+#       A solution will be evaluating the integrals for Sobol indices by a quadrature rule. Since
+#       we have a complete freedom, using GLL rule would be good. 
+import os
+import sys
+import numpy as np
+##import UQit.pce as pce
+import UQit.gpr_torch as gpr_torch
+import UQit.sobol as sobol
+import UQit.stats as statsUQit
+import UQit.reshaper as reshaper
+##import UQit.sampling as sampling
+#
+#
+class pSobol:
+    """
+    Probabilistic Sobol (pSobol) Sensitivity Indices over a p-D parameter space, 
+    where p=2,3,...
+    Model: y=f(q)+e
+
+    Args:
+      `qTrain`: 2D numpy array of shape (n,p)
+         Training samples for q
+      `yTrain`: 1D numpy array of size n
+         Training observed values for y
+      `noiseV`: 1D numpy array of size n
+         Standard-deviation of the noise in the training observations: e~N(0,noiseSdev)
+      `pSobolDict`: dict
+         Dictionary containing controllers for pSobol, including:
+           * 'qTest': List of length p
+              =[qTest_1,...,qTest_p], where qTest_i is a 1d numpy array of size nTest_i containing 
+              the uniformly-spaced test points in the i-th dimension
+           * 'pdf': List of length p
+              =[pdf_1,...,pdf_p], where pdf_i is a 1d numpy array of size nTest_i containing 
+              the PDF of q_i evaluated at qTest_i. 
+           * 'nMC': int
+              Number of independent samples drawn from the GPR to construct PCE 
+           * 'nIter_gpr': int
+              Number of iterations for optimization of the GPR hyper-parameters
+           * 'lr_gpr': float
+              Learning rate for optimization of the GPR hyper-parameters
+           * 'standardizeYTrain_gpr': bool (optional, default: False)
+              If true, the training data are standardized by shifting by mean and scaling by sdev:
+
+              :math:`yStnd =  (yTrain-mean(yTrain))/sdev(yTrain)`  
+
+           * 'convPlot_gpr': bool
+              If true, values of the hyper-parameters are plotted vs. iteration during the optimization.
+              
+    Attributes:
+      `fMean_list`: 1D numpy array of size `nMC`
+         PCE estimates for the mean of f(q)
+      `fVar_list`: 1D numpy array of size `nMC`
+         PCE estimates for the variance of f(q)
+      `optOut`: dict
+         Optional outputs for plotting using gprPlot, with the following keys:    
+            * `post_f`: Posterior density of f(q)
+            * `post_obs`: Posterior density of y
+            * `qTest`: A List of length p, 
[…]
