### 2021-03-12 eaf16d5
Fix xi -> q mapping in testSample
