+++ UQit/sampling.py
@@ -74,44 +74,44 @@ class trainSample:
 
     def check(self):
         if self.sampleType not in self.sampleTypeList:
-           raise KeyError('#ERROR @ parSample: Invalid sampleType! Choose from'\
+            raise KeyError('#ERROR @ parSample: Invalid sampleType! Choose from'\
                    ,self.sampleTypeList)
         if self.sampleType=='GQ' and self.GQdistType not in self.GQdistList:
-           raise KeyError('#ERROR @ parSample: Invalid GQdistType! Choose from'\
+            raise KeyError('#ERROR @ parSample: Invalid GQdistType! Choose from'\
                    ,self.GQdistList)
 
     def genSamples(self):       
         n=self.nSamp
         if self.sampleType=='GQ' and self.GQdistType in self.GQdistList:
-           self.gqPtsWts() 
+            self.gqPtsWts()
         elif self.sampleType=='normRand':
-           self.xi=np.random.normal(size=n)
-           self.xiBound=[min(self.xi),max(self.xi)]
-           self.mean=self.qInfo[0]
-           self.sdev=self.qInfo[1]
-           self.xi2q_map()         
+            self.xi=np.random.normal(size=n)
+            self.xiBound=[min(self.xi),max(self.xi)]
+            self.mean=self.qInfo[0]
+            self.sdev=self.qInfo[1]
+            self.xi2q_map()
         else:    
-           if self.sampleType=='GLL':
-              self.xiBound=[-1,1]
-              xi_,w_=nodes.gllPts(n) 
-              self.xi=xi_
-              self.w=w_
-           if self.sampleType=='unifSpaced':
-              xiBound_=[0,1]
-              self.xiBound=xiBound_
-              self.xi=np.linspace(xiBound_[0],xiBound_[1],n)
-           elif self.sampleType=='unifRand':
-              self.xiBound=[0,1]
-              self.xi=np.random.rand(n)
-           elif self.sampleType=='Clenshaw':
-              self.xiBound=[-1,1]
-              self.xi=nodes.Clenshaw_pts(n)
-           elif self.sampleType=='Clenshaw-Curtis':
-              self.xiBound=[0,1]
-              l_=1+int(mt.log(n-1)/mt.log(2))
-              self.xi=nodes.ClenshawCurtis_pts(l_)
-           self.qBound=self.qInfo
-           self.xi2q_map()
+            if self.sampleType=='GLL':
+                self.xiBound=[-1,1]
+                xi_,w_=nodes.gllPts(n)
+                self.xi=xi_
+                self.w=w_
+            if self.sampleType=='unifSpaced':
+                xiBound_=[0,1]
+                self.xiBound=xiBound_
+                self.xi=np.linspace(xiBound_[0],xiBound_[1],n)
+            elif self.sampleType=='unifRand':
+                self.xiBound=[0,1]
+                self.xi=np.random.rand(n)
+            elif self.sampleType=='Clenshaw':
+                self.xiBound=[-1,1]
+                self.xi=nodes.Clenshaw_pts(n)
+            elif self.sampleType=='Clenshaw-Curtis':
+                self.xiBound=[0,1]
+                l_=1+int(mt.log(n-1)/mt.log(2))
+                self.xi=nodes.ClenshawCurtis_pts(l_)
+            self.qBound=self.qInfo
+            self.xi2q_map()
 
     def gqPtsWts(self):
         """
@@ -119,18 +119,18 @@ class trainSample:
         """
         n=self.nSamp
         if self.GQdistType=='Unif':
[…]
