### 2021-03-16 #2871 @ctdegroot
@salrm8
This is a bit of an awkward statement in the paper: "It is important to note that uncertainty is different from error, which is defined as the deviation of a realization from a reference true value." While it is true that uncertainty is different than error, I think your definition of error is too limiting. For example, one may attempt to quantify discretization error while having no "true" value. Perhaps you may consider a definition where error is any recognizable error not due to lack of knowledge/information (https://www.grc.nasa.gov/www/wind/valid/tutorial/glossary.html#error). Otherwise I am satisfied with the conent of the paper.
