### 2021-03-16 #2871 @ctdegroot
@salrm8 I was unable to find instructions on how to run the tests or how to run examples. Please let me know how to do this? Ideally this should be on the main README page. Once these two items are handled, I am satisfied with my review.
