+++ UQit/psobol.py
@@ -4,14 +4,16 @@
 #--------------------------------------------------------------
 # Saleh Rezaeiravesh, salehr@kth.se
 #--------------------------------------------------------------
-# ToDo: For p>=3, choosing large number of test points to evaluate integrals in Sobol indices, 
-#       is problematic since associated samples are generated from GPR (GpyTorch has memory issues). 
-#       A solution will be evaluating the integrals for Sobol indices by a quadrature rule. Since
-#       we have a complete freedom, using GLL rule would be good. 
+# ToDo: For p>=3, choosing large number of test points to evaluate the 
+#  integrals in the Sobol indices can be a bit problematic since the 
+#  samples generated from GPR (GpyTorch) may lead to memory/CPU issues. 
+#  A solution  will be evaluating the integrals for Sobol indices by a 
+#  quadrature rule. Since we have a complete freedom, using GLL rule would be good. 
+#--------------------------------------------------------------
+#
 import os
 import sys
 import numpy as np
-##import UQit.pce as pce
 import UQit.gpr_torch as gpr_torch
 import UQit.sobol as sobol
 import UQit.stats as statsUQit
@@ -19,11 +21,11 @@ import UQit.reshaper as reshaper
 ##import UQit.sampling as sampling
 #
 #
-class pSobol:
+class psobol:
     """
-    Probabilistic Sobol (pSobol) Sensitivity Indices over a p-D parameter space, 
+    Probabilistic Sobol (psobol) Sensitivity Indices over a p-D parameter space, 
     where p=2,3,...
-    Model: y=f(q)+e
+    Surrogate model: y=f(q)+e
 
     Args:
       `qTrain`: 2D numpy array of shape (n,p)
@@ -32,8 +34,8 @@ class pSobol:
          Training observed values for y
       `noiseV`: 1D numpy array of size n
          Standard-deviation of the noise in the training observations: e~N(0,noiseSdev)
-      `pSobolDict`: dict
-         Dictionary containing controllers for pSobol, including:
+      `psobolDict`: dict
+         Dictionary containing controllers for psobol, including:
            * 'qTest': List of length p
               =[qTest_1,...,qTest_p], where qTest_i is a 1d numpy array of size nTest_i containing 
               the uniformly-spaced test points in the i-th dimension
@@ -55,24 +57,27 @@ class pSobol:
               If true, values of the hyper-parameters are plotted vs. iteration during the optimization.
               
     Attributes:
-      `fMean_list`: 1D numpy array of size `nMC`
-         PCE estimates for the mean of f(q)
-      `fVar_list`: 1D numpy array of size `nMC`
-         PCE estimates for the variance of f(q)
+      `Si_samps`: list of length p
+         =[S1_samps,S2_samps,...,Sp_samps] where Si_samps is a 1d numpy array of size nMC containing 
+         samples of the main Sobol index of the i-th parameter. 
+      `Sij_samps`: list of length (p-1)!
+         =[S12_samps,S13_samps,...] where Sij_samps is a 1d numpy array of size nMC containing samples 
+         of the dual interaction Sobol index between i-th and j-th parameters.
+      `STi_samps`: list of length (p-1)!
[…]
