### 2021-03-17 #2871 @salrm8
> @salrm8
> This is a bit of an awkward statement in the paper: "It is important to note that uncertainty is different from error, which is defined as the deviation of a realization from a reference true value." While it is true that uncertainty is different than error, I think your definition of error is too limiting. For example, one may attempt to quantify discretization error while having no "true" value. Perhaps you may consider a definition where error is any recognizable error not due to lack of knowledge/information (https://www.grc.nasa.gov/www/wind/valid/tutorial/glossary.html#error). Otherwise I am satisfied with the conent of the paper.

Thank you @ctdegroot for the comment. I understand the point of view, however, we are discussing the uncertainty in a QoI which is the outcome of a simulation, for instance. To emphasize this, we have modified the sentence  "It is important to note that uncertainty is different ..." as "It is important to note that **for a QoI**, uncertainty is different ...".

There are also a few ways to define error and uncertainty, for instance in AIAA Guide and UQ literature. But these definitions are not contradictory. Among other things, what is needed to be clearly specified is whether error/uncertainty in the modeling/codes or a QoI is addressed.

In the discretization error example by the reviewer,  if we use modified equation approach in finite-difference method for instance, then the solution around which the expansion is made can be considered as a reference truth. If instead we mean a posteriori error in QoIs, then we need to compare the numerical solution to continuous solution or experimental data: again we have a reference value.
However, there are cases where the combination of errors can lead to behaviors which could not be a-priori predicted or even be against what is expected. A good example is the scale-resolving simulation of turbulent flows and influence of the  numerical parameters at resolutions coarser than DNS. In this cases, the evolvement of errors may lead to uncertainties.
