### 2021-03-17 #2871 @salrm8
> @salrm8 I was unable to find instructions on how to run the tests or how to run examples. Please let me know how to do this? Ideally this should be on the main README page. Once these two items are handled, I am satisfied with my review.

Thank you @ctdegroot. We have added a short guide in README on how to run a test:

### How to run a test:
The tests associated to different UQ techniques are provided in `./tests/`.
Most of the tests have also been used in the notebooks and documentation.
For instance, to run a test from `pce_tests.py` in `./tests/`, do the following:

`python3 -c 'import pce_tests as X;X.pce_1d_test()'`
