### 2021-04-04 #2871 @ctdegroot
@salrm8 I accept your explanation of the reference value. However, I think it should not say "reference true value", but just "reference value". This would be more consistent with your own explanation with regards to a posteriori error estimations. Yes, it is a reference value as you explained, but I can't see how the word "true" applies in this case.

With regards to the change in the readme, explaining how to run a test, I am pleased.
