### 2021-04-04 #2871 @salrm8
@ctdegroot Thank you for your comment. What was meant in the text by "reference true data" is for instance, a reference DNS or experimental data for the quantities of interest obtained by minimizing the error and uncertainties from all possible sources. But we never get rid of the error and uncertainties completely. Still, the reported QoIs can be considered the most accurate values for describing the physics and hence be used for evaluation of other simulations/experiments.
In fact, from theoretical point of view, we have "true unobserved" values which can be estimated by numerical simulations and experiments.
