### 2021-04-14 30b95c5
Citation updates to JOSS paper
