### 2021-04-14 #2871 @kyleniemeyer
Hi @salrm8, I am the AEIC on duty this week, and it looks like this submission is essentially ready to be accepted. I noticed some small issues with citation formatting, and fixed them in https://github.com/KTH-Nek5000/UQit/pull/3. Can you merge this PR?

Once you've done that, you can archive your software repository (e.g., on Zenodo) and let us know the DOI here.
