### 2021-04-20 #2871 @salrm8
> @salrm8 ah, thanks - could you clean up the author list on the Zenodo archive to match the paper?

@kyleniemeyer Done! Thank you.
