### 2021-04-20 #2871 @salrm8
> Congrats @salrm8 on your article's publication in JOSS!
>
> Many thanks to @jayten and @ctdegroot for reviewing this submission, and @drvinceknight for editing.

Thank you @kyleniemeyer , @drvinceknight , @jayten, and @ctdegroot!
